Handle "hold" properly in tuya_1_to_4_buttons.yaml
Per https://github.com/Koenkk/zigbee2mqtt/issues/25772 z2m reports `<buttonNo>_<eventType>` for `single` and `double` events, but just `hold` for the long press. To determine which button is pressed you have to use `button` (and be sure to compare it to a numeric — not a string! — value). I added logging to see when the events are not processed correctly.